move test-utils to separate folder
