+++ react/src/views/Distributions/DistroSpotsView/components/DistroSpotsContainer.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { ComponentStory, ComponentMeta } from "@storybook/react";
 import { action } from "@storybook/addon-actions";
 import DistroSpotsContainer from "./DistroSpotsContainer";
-import { StorybookApolloProvider } from "utils/test-utils";
+import { StorybookApolloProvider } from "tests/test-utils";
 import {
   DistributionEventState,
   DistroSpotsForBaseIdQuery,
@@ -18,36 +18,36 @@ export default {
   decorators: [
     (Story) => {
       worker.use(
-        graphql.query<
-          DistroSpotsForBaseIdQuery,
-          DistroSpotsForBaseIdQueryVariables
-        >("DistroSpotsForBaseId", (req, res, ctx) => {
-          const mockedDistroSpotsForBaseIdData = {
-            base: {
-              __typename: "Base",
-              distributionSpots: [
-                {
-                  __typename: "DistributionSpot",
-                  id: "1",
-                  name: "Horgos (River)",
-                  latitude: 132.142,
-                  longitude: 132.142,
-                  distributionEvents: [
-                    {
-                      __typename: "DistributionEvent",
-                      id: "3",
-                      name: "Warm Clothes and Tea",
-                      // startDateTime: "2022-06-01T14:48:25+00:00",
-                      plannedStartDateTime: "2022-06-01T14:48:25+00:00",
-                      state: DistributionEventState.Planning,
-                    },
-                  ],
-                },
-              ],
-            },
-          } as DistroSpotsForBaseIdQuery;
-          return res(ctx.data(mockedDistroSpotsForBaseIdData));
-        })
+        graphql.query<DistroSpotsForBaseIdQuery, DistroSpotsForBaseIdQueryVariables>(
+          "DistroSpotsForBaseId",
+          (req, res, ctx) => {
+            const mockedDistroSpotsForBaseIdData = {
+              base: {
+                __typename: "Base",
+                distributionSpots: [
+                  {
+                    __typename: "DistributionSpot",
+                    id: "1",
+                    name: "Horgos (River)",
+                    latitude: 132.142,
+                    longitude: 132.142,
+                    distributionEvents: [
+                      {
+                        __typename: "DistributionEvent",
+                        id: "3",
+                        name: "Warm Clothes and Tea",
+                        // startDateTime: "2022-06-01T14:48:25+00:00",
+                        plannedStartDateTime: "2022-06-01T14:48:25+00:00",
+                        state: DistributionEventState.Planning,
+                      },
+                    ],
+                  },
+                ],
+              },
+            } as DistroSpotsForBaseIdQuery;
+            return res(ctx.data(mockedDistroSpotsForBaseIdData));
+          },
+        ),
       );
       return (
         <StorybookApolloProvider>
